Changed linear() return sign.
Maybe I'm not understandign the reason behind this choiche, but I noticed that this function was used only once in the trace_into_index() function
with opposite sign. Because this represent a semantic exception between al the other cases in trace_into_index(), I suggest to change the sign in both places.
Nothing should change, but this will ease me to use this function without a special treatment later, when I will insert the code for autoscale.