Improved scale format in LOGMAG mode
Another preliminary change for the autoscale feature: dB/ will be shown as int above 10 dB/ and as X.X when below.
I've saved the current value of the scale in a local variable within trace_get_info() in order to not make a double
call to get_trace_scale() in the TRC_LOGMAG case, so I replaced all the other calls with the local variable for
the same reason.